requirements.txt
PyQt6
numpy
Pillow
